+++ Bubble/Modules/HorseDetailPage/PageVM/NewHouseDetailPageViewModel.swift
@@ -27,7 +27,7 @@ class NewHouseDetailPageViewModel: NSObject, DetailPageViewModel {
 
     }
 
-    func requestData(houseId: Int) {
+    func requestData(houseId: Int64) {
         requestNewHouseDetail(houseId: 6573911052528910605)
                 .debug()
                 .subscribe(onNext: { [unowned self] (response) in
